test: improve test suite organization and clarity
Reorganize middleware tests from Unit to Feature directory as they require full Laravel application context. Remove nested describe() blocks in favor of flat structure with self-explanatory test titles. Consolidate performance tests into main feature test file to reduce duplication.
