+++ tests/Pest.php
@@ -15,4 +15,7 @@ use Mirror\Tests\TestCase;
 
 pest()->extend(TestCase::class)
     ->use(Illuminate\Foundation\Testing\RefreshDatabase::class)
+    ->in('Feature');
+
+pest()->extend(TestCase::class)
     ->in('Unit');
