+++ workbench/app/Models/User.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace Workbench\App\Models;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Mirror\Concerns\Impersonatable;
-use Workbench\Database\Factories\UserFactory;
 
 class User extends Authenticatable
 {
@@ -44,13 +43,5 @@ class User extends Authenticatable
             'email_verified_at' => 'datetime',
             'password' => 'hashed',
         ];
-    }
-
-    /**
-     * Create a new factory instance for the model.
-     */
-    protected static function newFactory(): UserFactory
-    {
-        return UserFactory::new();
     }
 }
