+++ workbench/database/factories/UserFactory.php
@@ -1,52 +1,21 @@
 <?php
 
-namespace Workbench\Database\Factories;
+namespace Database\Factories;
 
+use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Str;
-use Workbench\App\Models\User;
 
 /**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\Workbench\App\Models\User>
+ * @template TModel of User
+ *
+ * @extends Factory<TModel>
  */
-class UserFactory extends Factory
+class UserFactory extends \Orchestra\Testbench\Factories\UserFactory
 {
     /**
      * The name of the factory's corresponding model.
      *
-     * @var class-string<\Illuminate\Database\Eloquent\Model>
+     * @var class-string<TModel>
      */
     protected $model = User::class;
-
-    /**
-     * The current password being used by the factory.
-     */
-    protected static ?string $password = null;
-
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
-    public function definition(): array
-    {
-        return [
-            'name' => fake()->name(),
-            'email' => fake()->unique()->safeEmail(),
-            'email_verified_at' => now(),
-            'password' => static::$password ??= Hash::make('password'),
-            'remember_token' => Str::random(10),
-        ];
-    }
-
-    /**
-     * Indicate that the model's email address should be unverified.
-     */
-    public function unverified(): static
-    {
-        return $this->state(fn (array $attributes) => [
-            'email_verified_at' => null,
-        ]);
-    }
 }
